+++ sakura_core/view/figures/CFigureStrategy.h
@@ -17,7 +17,7 @@ class CFigure{
 public:
 	virtual ~CFigure(){}
 	virtual bool DrawImp(SColorStrategyInfo* pInfo) = 0;
-	virtual bool Match(const wchar_t* pText, int nTextLen) const = 0;
+	virtual bool Match(const wchar_t* pText, ssize_t nTextLen) const = 0;
 
 	//! 色分け表示対象判定
 	virtual bool Disp(void) const = 0;
@@ -49,7 +49,7 @@ public:
 	static int FowardChars(SColorStrategyInfo* pInfo);
 
 	bool DrawImp(SColorStrategyInfo* pInfo) override;
-	bool Match([[maybe_unused]] const wchar_t* pText, [[maybe_unused]] int nTextLen) const override
+	bool Match([[maybe_unused]] const wchar_t* pText, [[maybe_unused]] ssize_t nTextLen) const override
 	{
 		return true;
 	}
