+++ sakura_core/view/figures/CFigure_HanSpace.h
@@ -15,7 +15,7 @@
 class CFigure_HanSpace final : public CFigureSpace{
 public:
 	//traits
-	bool Match(const wchar_t* pText, int nTextLen) const override;
+	bool Match(const wchar_t* pText, ssize_t nTextLen) const override;
 
 	//action
 	void DispSpace(CGraphics& gr, DispPos* pDispPos, CEditView* pcView, bool trans) const override;
